Added trigger v4l2 controls
The alvium driver now supports triggering via v4l2 controls. The added v4l2 controls have the same ids as in the Nvidia alvium driver.
The following controls were added:
- V4L2_CID_TRIGGER_MODE: Boolean control to enable or disable the trigger mode.
- V4L2_CID_TRIGGER_ACTIVATION: Menu control to selection trigger activation type.
- V4L2_CID_TRIGGER_SOURCE: Menu control for selecting the trigger source.
- V4L2_CID_TRIGGER_SOFTWARE: Button control to send a software trigger signal.

For adding the trigger controls the v4l2 control handling in the driver was extended to support custom controls and the changes are also a preparation for a more generic control handling implementation. These changes are also including flags for disabling and enabling controls while streaming. If the flag AVT_CTRL_STREAM_DISABLED is set, then the grabbed flag of the control is set, when the streaming is started. Similar removes the flag AVT_CTRL_STREAM_ENABLED the grabbed flag on the stream start.