Fixed some issues with the v4l2 controls and framerate enumeration
The bcrm_regmap_write functions were returning a positive value instead 0 when the register write was successful and 0 instead of a negative error code when it failed. This behaviour is corrected now. Additionally, error handling for some regmap calls were added.
Besides an issue that occurred due to an incorrect implementation in enum_frame_intervals function is fixed and now frame intervals that have a numerator other than 1 or 1000 are supported.