Fixed kernel panic when the driver is loaded as module on the Xilinx ZCU106
When the driver was loaded as kernel module on a Xilinx ZCU106, a kernel panic occurred during the completion of the subdevice registration. This happend, because the control handler of the subdevice was not initialized and therefore the pointer to the mutex of the control handler was zero. The initialization order is fixed now.
This bug does not occur when the driver was build in, because in this case the camera driver is initialized before the platform driver and therefore the subdevice registration is not completed immediately. On the i.MX8M Plus the camera driver could be load successfully as module before, but only because the platform drivers are loaded as modules to and then are initialized after the camera driver.