Use bcrm_regmap_write64 instead of regmap_bulk_write when writing 64bit registers.
Previously regmap_bulk_write was used for writing 64bit registers of v4l2 controls. Because regmap_bulk_write is a standard kernel function and not using the bcrm write handshake, the v4l2 set control ioctls succeeded before the write request was completely process by the camera. The function bcrm_regmap_write64, that is now used, waits for write handshake to be finished before it returns.