Rework avt3_detect to work correctly on the imx8mp
The old version avt3_detect was just performing a dummy write and read to detect if a camera is connected. This was not working correctly on the imx8mp, so now the function performs a complete register read of the cci layout version register (address 0x0000). This register is used, because the endianness of the register address is not relevant for this address. In the end it is only checked if anything was read from the camera. The check if the cci layout version is compatible with the driver version is performed later during the probe. The only purpose of avt3_detect is to fast check if a camera is connected.