Enabled treat warnings as errors for the driver build and fixed found issues
When the driver is build now all warnings are treat as errors (-Werror) to improve the quality of the driver source code and avoid issues. This is also helpful when the driver is build using the yocto meta layer, because bitbake normally hides the compiler warnings. All issues that were found due to using Werror are fixed. The issues were mostly unused and outdated code.