Added support for digital binning and updated cropping implementation
The binning can be selected via the format ioctls or the selection api using the compose target. The user can request any frame size. The requested frame size will be adjusted to the closest matching binning setting. The implementation follows the v4l2 standard for the selection api, so the cropping is applied before the binning. If the user changes the roi using the crop selection target, the binning is reset to 1x1 and frame size is the size accordingly. The limit for the cropping is always the complete sensor.
 Some v4l2 ioctls that are only valid for video device are removed now.
 Added .clang-format file from the linux kernel to ensure the driver source code is formatted correctly.