Fix CSIGENERAL-2734: Cameras with large sensor are not detected
When cameras with large sensors were used the driver load time was too high because of the binning frame size readout and no video device was registered. Now the binning frame sizes are no longer readout from the camera but calculated within the driver. This reduces the load time.
The driver prints now the camera name and firmware version to the kernel log when it is loaded.