Fix CSIGENERAL-2732: low framerates on imx8mp
The driver now supports the auto framerate feature of the camera. The auto framerate is enabled by setting the frame interval to a zero value.
Enum frame intervals now only returns one value, which is the maximum supported framerate.
On stream start the framerate is now automatically adjusted to match the camera limits so the stream can always be started.
Removed some outdated framerate code.