Fix auto white balance not working
The auto white balance v4l2 control was only working once after it was enabled. If e.g. the red balance value was changed while auto white balance was enabled, the white balance was not adjusted again. The problem was that if the auto white balance v4l2 control was enabled the value 1 was written into the register, but the corrected value is 2, because 1 means auto white balance once and not continuous.