Return framesizes per supported binning setting
The enum framesizes call in the driver returns foreach supported binning setting the corresponding framesize instead of only one continuous framesize. The framesizes are calculated respecting the current cropping. So if the cropping is set to 1024x1024, for 2x2 a framesize of 512x512 and for 4x4 256x256 and so on will be returned. This makes it possible to directly see the resulting framesizes after applying the binning.