Added sensor binning
The driver now supports the configuration of the sensor binning. Therefore, the v4l2 control binning_selector was added. The control is a menu control and supports the values "Digital" and "Sensor". If the binning selector control is changed, the driver tries to match the previous binning setting. So if previously 2x2 digital binning was used, the driver tries to enable 2x2 sensor binning. If the frame size is change by selecting another binning type, a source change event is generated.
In firmware version without the sensor binning changes the byte order of the binning inquiry register is swapped. To make the driver compatible with all firmware version a workaround was implemented that checks if an inquiry bit outside the defined range is set and then switches the byte order.