Remove platform abstraction layer
Removed the platform abstraction layer, because it is not needed in this form and added too much overhead.
From now on the "camera_common_data" that is required by the NVIDIA drivers is directly stored in the driver data struct, if the NVIDIA define is set. Also only one compatible string "alliedvision,avt_csi2" is now available. This also prevents users from using the wrong compatible string and then accidentally bricking their system.