Added support for Orin Nano/NX modules in super mode
When in super mode the orin nano and nx modules use a different compatible string with the suffix "-super" compared to the default mode. The new compatible strings are now added to the orin nano / nx device tree overlay, so the driver can be used also on system in super mode.