Added test pattern v4l2 control
The driver now exposes the test pattern feature of the firmware 15 as v4l2 control. For this the predefined v4l2 control V4L2_CID_TEST_PATTERN is used. Also the custom control defintion AVT_CID_TEST_PATTERN was removed.
Which test patterns are available is queried during initialization and written to the skip mask field of the test pattern control.