Added v4l2 control for FW15 user data bcrm feature
The new user data firmware features is exposed as single v4l2 control. The control is a u32 array control with an length of 7. This gives the user the most flexibility to store any data in the camera user data space. The driver only writes the changed values to the camera.