Refactor io line handling
Added new io line handling functionality to avoid / block multiple usage of the same io line (e.g. expsure active and triggering at the same time). If a feature / v4l2 controls is using a io line it has to call "avt_line_get" before feature is activated in the camera. "avt_line_get" configures the io line (input / output, (non) inverted). If the io line is already in use -EBUSY is returned. When a feature that is using a io line is disabled it must call "avt_line_put" to free the io line, so it can be used by other features. "avt_line_put" always configures the line as an input to avoid any unwanted behaviour.