Prepare for new power management implementation
As perperation for a new power management implementation the old s_power functions aswell as the hardreset were removed. The camera does not have a reset gpio and the power down functionality should use the regulator api instead.
The softreset implementation is reworked and now uses read_poll_timeout instead the custom polling, which results in a smaller code size.