First version of regulator based power managment
It is now possible to specify a regulator that powers the Alvium camera using the "vcc-ext-supply" device tree property. Currently the regulator is always enabled on driver probe and only disabled on driver remove.
During initialization the driver now waits up to 10s for a camera to boot up, because it might be that it was just powered on using the regulator. If no regulator was specified the driver performs the softreset as the camera seems to be externaly powered and we don't known in which state it is.
Reworked the detect function to use the standard register access.