Fixed default value read of boolean controls trigger mode and frame trigger wait line mode
The __v4l2_ctrl_s_ctrl in the privious version was failing, because not at all controls were fully in initilized and therefore the value was not changed. The new version now directly stores the value in the ctrl struct. Also for all depending controls the grab status is set correctly.