Add intensity controler region control
The driver contains now the menu control "intensity controller region" it can be used to change the image region on which the intensity controller is operating. The currently available options are "Full Image" and "Auto Region". If the "intensity controller region" is set "Auto Region" the intensity controller will use the image area specified in the auto region controls