Disable mode switch in driver inititalisation
The driver performed always an switch into bcrm mode in initialization, even when the camera was already in bcrm mode. This causes issue with the firmware as some parameters are reset on the mode switch. Instead of performing the mode switch the drivers reads now the current mode from the camera and aborts the initialisation if the camera is not in bcrm mode.