Fixed probe fail if vcc-ext regulator is no present
If the vcc-ext property was not present in the device tree, the driver probe failed, because devm_regulator_get_optional returns an error code if the regulator is not available. This is now fixed by only enabling and setting the reg_vcc_ext in the private data struct, when the regulator is present.