Fixed framerate setting not applied to gstreamer pipeline
If a fixed framerate is requested, the v4l2src does first check if this framerate is available and fails when it is not. The Alvium subdevice driver always returned the maxium framerate in enum frame interval. Therefore only the maximum framerate can be used with the v4l2src.
One limitation here is that the v4l2 subdev api does only support discrete framerates and not the continuous or steps options as the video device api.  To fix the issue the Alvium driver now returns two discrete framerates:
    0 -> Current framerate
    1 -> Maximum framerate
With that it is now possible to use any other framerate then the maximum with the v4l2src, but the framerate must be set beforehand using the v4l2 subdev api.