Add v4l2 flash support
It is now possible to specify a connected v4l2 flash light using the "flash" device tree property. A v4l2 async subdev notifier is used to resolve the device tree reference and to create the /dev/v4l-subdevX node of the flash. The notifier bound callback stores the given flash v4l2 subdevice in the "flash_sd" field of the driver data structure. On s_stream the driver will call the s_stream of the flash, if a flash is bound.