Add dummy implementation for s_power
Add dummy implementation, which just returns 0, for the s_power core operation. Even though s_power is deprecated in the mainline kernel. Some drivers still rely on this operation to be present. E.g. the imx8-isi driver streamon fails if the s_power operation of the camera/sensor subdevice does not succeed.