Fix v4l2 control "serial_number" not set correctly
The issue was introduced with the commit "On return from power save mode the time is measured and printed to the kernel log [f3863e82019ca9411d49b76cf0e57a4fd69b68d6]", because it changed the behaviour of the "__set_special_ctrl" to return -ENOTTY, if an not handelt control is set. Due to this change "__v4l2_ctrl_s_ctrl_string" in the initialization of the serial numer control is always failing. The issue is fixed by directly copying the current value into the "p_cur" field of the control struct using "strscpy". Also calling "__v4l2_ctrl_s_ctrl_string" is not needed for the string controls as they are read only and just need to be initialized to the correct value.