seperate bcm and gencp format handling
When setting the format in gencp mode, the current rect size (crop) was always set to the size of GenICam data format. Crop settings are not correctly reset when switching back to bcm mode leading format missmatch issue on stream start. To overcome this issue the format handling is completly seperated and depending on the mode different field private driver data is used. Whenever the format is accessed the new helper "avt_get_mode_fmt", which returns correct format pointer for the current mode, shall be used.