Revert camera subdevice multiple open
The commit "35fcabedbaeeb1b4efcd2dc7538dacefdf961cba" also contained changes that allowed multiple opens of the camera subdevice. This was not intended to be merged as the changes were not fully tested and might introduce some side effects. So, they are reverted with this commit. Also some whitespace issues from the PR review are fixed.