Fully support the Alvium power save mode
In commit d3a61394d48500fbe138bcb83209fd9ac0768e2d the v4l2 control for the Alvium power save mode was added. But as the Alvium camera resets all register settings when entering the power safe mode some extra handling needs to be implement. The driver now supports to different behaivours when exiting the power save mode.
The first and default one writes all current settings from the driver into the camera. This restores the register values to the state they where when the power save mode was entered. If a v4l2 control is modifed while the camera is in power save mode, the new value is also correctly written to the camera once the power save mode is deactived.
The second resets all all v4l2 controls to their default values. This behaviour can be activated by setting the module parameter "power_save_reset_controls" to "Y".