Adjust reverse x/y handling for FW15
Starting with FW15 the width and height of the image are adjusted when using reverse x and y to always provide the same bayer pattern. When the reverse x or y v4l2 control is set the driver reads the new maximum width / height from the camera and adjusts the current format accordingly. The new driver behaivour is only activated when the BCRM version >= 1.28, which is the BCRM version used in FW15.