Add v4l2 controls for sensor identifiaction register
The sensor identifiaction register is added in BCRM version 1.28 and contains information about the sensor used in the Alvium. For each of fields "Sensor Id", "Sensor Platform Id", and "Sensor Info" a v4l2 control is added. The check bits information of the sensor identification register is included in the log status message. Addiontally, the log status function was reworked to print "n/a", when the underlying feature is not implemened in the camera firmware.