Fix auto region control limit updates
In c8c182c806e0ef133ccefa4968604fd7a79e853a the way how the limits of the auto region controls are update was changed to calculate the values inside the driver instead of reading them from the registers. But the implementation was not done correctly as for the width and height the minimum was changed and also the new limits were not calculated correctly. When ever the value of a auto region control is updated the *maximum* value of the corresponding control must be updated to the ROI width / height - the new control value. E.g. when the auto region width is changed, the maximum of the offset x must be set to the current to "ROI width - auto region width". Also whenever the ROI is changed the auto region controls must be updated. When the ROI is changed the auto region is now reset to the new ROI.