Fix kernel panic when driver is not correctly build for NVIDIA systems
The driver causes a kernel panic, when it is used on NVIDIA system, but was not built with NVIDIA support. This happens, because the NVIDIA drivers assume that the camera_common structure is used, when a modeX node is present in the device tree. To avoid the crash the driver checks now during initialization, if the device tree contains a "modeX" node, and fails, when the driver was built without NVIDIA support and the node is present.