Fix kernel panic when device tree nodes are missing
When the "endpoint" node in the device tree does not contain the properties "clock-lanes", "data-lanes", and "link-frequencies" the driver causes a kernel panic during initialization. To fix the issue the "v4l2_fwnode_endpoint" is now zero inialized to avoid crashes due to random values. Addionally the test if a given lane count is supported by the camera is reworked to avoid any issue due to overflows. With the current implementation a lane count of zero causes a kernel panic, because 1 is substraced from the lane count, which results in -1, and than "test_bit" performes an invalid access.