Implement get_frame_desc pad operation
A function for the get_frame_desc pad operation is provided but it does not contain any implementation and just returns 0. As the PiSP Camera Front End driver relies on the operation it is now implemented correctly. The number of entries in the frame description is hardcoded to one, because a Alvium camera does only support one fixed format setting.