Rework device tree overlay structure and add raspberry pi5 overlay
The device tree overlay directory is restructured to support overlay for multiple platforms. The existing device tree overlays for NVIDIA Jetson boards moved to the nvidia subdirectory. For the new Raspberry PI 5 overlay the raspberrypi subdirectory is added. Each subdirectory conatins it's own Kbuild file in which the overlays to build are specified. Also the overlay installation process is updated and now recursivly searchs for .dtbo files in the overlay folder. With this it is no longer required to add the overlay file in two places.