Fix driver can't be build for NVIDIA systems
After the merge of PR #49 the driver is no longer building for NVIDIA systems, because the function "has_jetson_nodes" is not used. The issue is fixed by using the function to print a warning message, when the driver is build with NVIDIA support, but no "modeX" device tree node is present.